Convert from module to package and update test processing.
