+++ .gitignore
@@ -1,5 +1,32 @@
 
-# Created by https://www.gitignore.io/api/pydev,python,pycharm
+# Created by https://www.gitignore.io/api/osx,pydev,python,pycharm,visualstudiocode
+
+### OSX ###
+*.DS_Store
+.AppleDouble
+.LSOverride
+
+# Icon must end with two \r
+Icon
+
+# Thumbnails
+._*
+
+# Files that might appear in the root of a volume
+.DocumentRevisions-V100
+.fseventsd
+.Spotlight-V100
+.TemporaryItems
+.Trashes
+.VolumeIcon.icns
+.com.apple.timemachine.donotpresent
+
+# Directories potentially created on remote AFP share
+.AppleDB
+.AppleDesktop
+Network Trash Folder
+Temporary Items
+.apdisk
 
 ### PyCharm ###
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and Webstorm
@@ -172,5 +199,13 @@ venv.bak/
 # mypy
 .mypy_cache/
 
+### VisualStudioCode ###
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+.history
 
-# End of https://www.gitignore.io/api/pydev,python,pycharm
+
+# End of https://www.gitignore.io/api/osx,pydev,python,pycharm,visualstudiocode
