+++ tox.ini
@@ -6,11 +6,14 @@ minversion = 1.6.0
 deps =
     pytest
     flake8
-    import-order >= 0.0.10
+    flake8-import-order
 commands =
-    py.test test.py
-    flake8 .
-    import-order --only-file sqlalchemy_utc.py test.py
+    pytest tests
+    flake8
 
 [flake8]
-filename = sqlalchemy_utc.py,test.py
+max-complexity = 10
+exclude =
+    .git,
+    .tox,
+    __pycache__
